pyproject.toml
[project]
name = "crucible-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "coverage>=7.13.5",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.1",
    "h5py>=3.16.0",
    "hyperspy>=2.4.0",
    "igor2>=0.5.13",
    "imageio>=2.37.3",
    "matplotlib>=3.10.8",
    "nano-crucible>=2.1.0",
    "ncempy>=1.15",
    "pika>=1.3.2",
    "pillow>=12.1.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "pyyaml>=6.0.3",
    "xmltodict>=1.0.4",
]
